minor fixes

+++ src/main/java/com/example/web2/Model.java
@@ -47,15 +47,15 @@ public class Model {
     }
 
     private static boolean checkTriangle(double x, double y, double r) {
-        return x >= 0 && y <= 0 && y - ((double) x) >= -r;
+        return x >= 0 && y <= 0 && y - x >= -r;
     }
 
     private static boolean checkRectangle(double x, double y, double r) {
-        return x >= 0 && y >= 0 && ((double) x) <= r && y <= r;
+        return x >= 0 && y >= 0 && x <= r && y <= r;
     }
 
     private static boolean checkCircle(double x, double y, double r) {
-        return x <= 0 && y >= 0 && ((double) (x * x)) + y*y <= r*r/4;
+        return x <= 0 && y >= 0 && x * x + y*y <= r*r/4;
     }
 
     private static boolean checkHit(double x, double y, double r) {
